// src/app/share/model/product-search-detail.ts
export interface ProductSearchDetail {
  createDate: Date;

  description: string;

  id: number;

  idCategory: number;

  idSale: number;

  idMark: number;

  imageProduct: string;

  maSP: string;

  nameSP: string;

  price: number;

  sellCount: number;

  totalItem: number;

  idColor: number;

  codeSale: number;
}